fix: fix error could not extract column [6] from JDBC

+++ microservice-pos/src/main/java/com/cumpleanos/pos/persistence/repository/ReciboPOSViewRepositorio.java
@@ -3,10 +3,9 @@ package com.cumpleanos.pos.persistence.repository;
 import com.cumpleanos.pos.persistence.entity.ReciboPOSView;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface ReciboPOSViewRepositorio extends JpaRepository<ReciboPOSView, BigInteger> {
+public interface ReciboPOSViewRepositorio extends JpaRepository<ReciboPOSView, Long> {
 
     Optional<ReciboPOSView> findByAlmacenAndPventa(Long almacen, Long pventa);
     Optional<ReciboPOSView> findByUsrLiquidaAndEmpresa(Long usrLiquida, Long empresa);
